Former-commit-id: d4463e7e4f4b3e3419a45fb236e160512afb267b

+++ MathGamesAPI/app/controllers/notification.controller.js
@@ -13,17 +13,18 @@ exports.create = (req, res) => {
     return;
   }
 
-  /*
-  let date = new Date();
-  date.setTime( date.getTime() - new Date().getTimezoneOffset()*60*1000 );
-  date = date.toISOString().slice(0, 19).replace('T', ' ');
-  */
+  if (req.body.sender !== req.userId) {
+    res.status(401).send({
+      message: "Unauthorized!"
+    });
+    return;
+  }
+
   // Create a Notification
   const notification = {
     sender: req.body.sender,
     receiver: req.body.receiver,
     notification_type: req.body.notification_type
-    //notification_date: date
   };
 
   if (notification.notification_type === "F") {
@@ -84,6 +85,13 @@ exports.findByUserId = (req, res) => {
   // SELECT Notifications.id, User.id as sender_id, username as sender, receiver, notification_type, notification_date 
   // FROM Notifications JOIN User ON sender = User.id WHERE receiver = ? ORDER BY notification_date desc
 
+  if (parseInt(req.userId) !== parseInt(id)) {
+    res.status(401).send({
+      message: "Unauthorized!"
+    });
+    return;
+  }
+
   Notification.findAll({attributes: ['id', 'receiver', 'notification_type', 'createdAt' ], 
                         where: {receiver: id}, include: [{model: User, as: 'sender_user', attributes: [['id', 'sender_id'], ['username', 'sender']], 
                         order: [["createdAt", 'DESC']] }]})
@@ -102,25 +110,39 @@ exports.findByUserId = (req, res) => {
 exports.delete = (req, res) => {
   const id = req.params.id;
 
-  Notification.destroy({
-    where: { id: id }
-  })
-    .then(num => {
-      if (num == 1) {
-        res.send({
-          message: "Notification was deleted successfully!"
-        });
-      } else {
-        res.send({
-          message: `Cannot delete Notification with id=${id}. Maybe Notification was not found!`
-        });
-      }
-    })
-    .catch(err => {
-      res.status(500).send({
-        message: "Could not delete Notification with id=" + id
+  Notification.findOne({where: {id: id}}).then(notif => {
+    if (parseInt(notif.receiver) !== parseInt(req.userId)) {
+      res.status(401).send({
+        message: "Unauthorized!"
       });
-    });
+      return;
+    } else {
+      Notification.destroy({
+        where: { id: id }
+      })
+        .then(num => {
+          if (num == 1) {
+            res.send({
+              message: "Notification was deleted successfully!"
+            });
+          } else {
+            res.send({
+              message: `Cannot delete Notification with id=${id}. Maybe Notification was not found!`
+            });
+          }
+        })
+        .catch(err => {
+          res.status(500).send({
+            message: "Could not delete Notification with id=" + id
+          });
+        });
+
+    }
+  }).catch(err => {
+    res.status(500).send({
+        message: "Notification not found."
+    })
+});
 };
 
 // Delete all Notifications from the database.
